CIV-7811: Add OpenApi Publisher test to publish the docs
