+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/config/OpenAPIPublisherTest.java
@@ -1,0 +1,66 @@
+package uk.gov.hmcts.reform.civil.config;
+
+import com.microsoft.applicationinsights.web.internal.WebRequestTrackingFilter;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.mock.web.MockFilterConfig;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.web.context.WebApplicationContext;
+
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
+
+/**
+ * Built-in feature which saves service's Open Api specs in temporary directory.
+ * Each travis run on master should automatically save and upload (if updated) documentation.
+ */
+
+@SpringJUnitWebConfig
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles("integration-test")
+class OpenAPIPublisherTest {
+
+    @Autowired
+    private MockMvc mvc;
+
+    @Autowired
+    private WebApplicationContext wac;
+
+    @BeforeEach
+    void setUp() {
+        WebRequestTrackingFilter filter;
+        filter = new WebRequestTrackingFilter();
+        filter.init(new MockFilterConfig());
+        mvc = webAppContextSetup(wac).addFilters(filter).build();
+    }
+
+    @DisplayName("Generate Open API documentation")
+    @Test
+    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+    void generateDocs() throws Exception {
+        byte[] specs = mvc.perform(get("/v3/api-docs"))
+            .andExpect(status().isOk())
+            .andReturn()
+            .getResponse()
+            .getContentAsByteArray();
+
+        if (!Files.exists(Paths.get("/tmp/"))) {
+            Files.createDirectories(Paths.get("/tmp/"));    // needed on Windows systems as the directory won't exist by default
+        }
+        try (OutputStream outputStream = Files.newOutputStream(Paths.get("/tmp/openapi-specs.json"))) {
+            outputStream.write(specs);
+        }
+    }
+}
