--- src/integrationTest/java/uk/gov/hmcts/reform/civil/config/SwaggerPublisherTest.java
@@ -1,38 +1,0 @@
-package uk.gov.hmcts.reform.civil.config;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import uk.gov.hmcts.reform.civil.controllers.BaseIntegrationTest;
-
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-/**
- * Built-in feature which saves service's swagger specs in temporary directory.
- * Each travis run on master should automatically save and upload (if updated) documentation.
- */
-class SwaggerPublisherTest extends BaseIntegrationTest {
-
-    @DisplayName("Generate swagger documentation")
-    @Test
-    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
-    void generateDocs() throws Exception {
-        byte[] specs = mockMvc.perform(get("/v2/api-docs"))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsByteArray();
-
-        if (!Files.exists(Paths.get("/tmp/"))) {
-            Files.createDirectories(Paths.get("/tmp/"));    // needed on Windows systems as the directory won't exist by default
-        }
-        try (OutputStream outputStream = Files.newOutputStream(Paths.get("/tmp/swagger-specs.json"))) {
-            outputStream.write(specs);
-        }
-
-    }
-}
